+++ packages/react/src/Checkbox/Checkbox.test.tsx
@@ -2,7 +2,8 @@ import {render} from '@testing-library/react'
 import userEvent from '@testing-library/user-event'
 import React from 'react'
 import Checkbox from '../Checkbox'
-import {behavesAsComponent, checkExports, expectRendersWithClassname} from '../utils/testing'
+import {behavesAsComponent, checkExports} from '../utils/testing'
+import {FeatureFlags} from '../FeatureFlags'
 
 describe('Checkbox', () => {
   beforeEach(() => {
@@ -15,8 +16,22 @@ describe('Checkbox', () => {
   })
 
   it('should support `className` on the outermost element', () => {
-    const element = <Checkbox className={'test-class-name'} />
-    expectRendersWithClassname(element, 'test-class-name')
+    const Element = () => <Checkbox className={'test-class-name'} />
+    const FeatureFlagElement = () => {
+      return (
+        <FeatureFlags
+          flags={{
+            primer_react_css_modules_team: true,
+            primer_react_css_modules_staff: true,
+            primer_react_css_modules_ga: true,
+          }}
+        >
+          <Element />
+        </FeatureFlags>
+      )
+    }
+    expect(render(<Element />).container.firstChild).toHaveClass('test-class-name')
+    expect(render(<FeatureFlagElement />).container.firstChild).toHaveClass('test-class-name')
   })
 
   it('renders a valid checkbox input', () => {
