+++ packages/react/src/CounterLabel/CounterLabel.test.tsx
@@ -1,8 +1,9 @@
 import React from 'react'
 import {CounterLabel} from '..'
-import {behavesAsComponent, checkExports, expectRendersWithClassname} from '../utils/testing'
+import {behavesAsComponent, checkExports} from '../utils/testing'
 import {render as HTMLRender} from '@testing-library/react'
 import axe from 'axe-core'
+import {FeatureFlags} from '../FeatureFlags'
 
 describe('CounterLabel', () => {
   behavesAsComponent({Component: CounterLabel, options: {skipAs: true, skipSx: true}})
@@ -12,8 +13,22 @@ describe('CounterLabel', () => {
   })
 
   it('should support `className` on the outermost element', () => {
-    const element = <CounterLabel className={'test-class-name'} />
-    expectRendersWithClassname(element, 'test-class-name')
+    const Element = () => <CounterLabel className={'test-class-name'} />
+    const FeatureFlagElement = () => {
+      return (
+        <FeatureFlags
+          flags={{
+            primer_react_css_modules_team: true,
+            primer_react_css_modules_staff: true,
+            primer_react_css_modules_ga: true,
+          }}
+        >
+          <Element />
+        </FeatureFlags>
+      )
+    }
+    expect(HTMLRender(<Element />).container.firstChild).toHaveClass('test-class-name')
+    expect(HTMLRender(<FeatureFlagElement />).container.firstChild).toHaveClass('test-class-name')
   })
 
   it('renders a <span>', () => {
