+++ packages/react/src/Link/__tests__/Link.test.tsx
@@ -1,8 +1,9 @@
 import React from 'react'
 import Link from '..'
-import {render, behavesAsComponent, checkExports, expectRendersWithClassname} from '../../utils/testing'
+import {render, behavesAsComponent, checkExports} from '../../utils/testing'
 import {render as HTMLRender} from '@testing-library/react'
 import axe from 'axe-core'
+import {FeatureFlags} from '../../FeatureFlags'
 
 describe('Link', () => {
   behavesAsComponent({Component: Link})
@@ -12,8 +13,22 @@ describe('Link', () => {
   })
 
   it('should support `className` on the outermost element', () => {
-    const element = <Link href="#" className={'test-class-name'} />
-    expectRendersWithClassname(element, 'test-class-name')
+    const Element = () => <Link href="#" className={'test-class-name'} />
+    const FeatureFlagElement = () => {
+      return (
+        <FeatureFlags
+          flags={{
+            primer_react_css_modules_team: true,
+            primer_react_css_modules_staff: true,
+            primer_react_css_modules_ga: true,
+          }}
+        >
+          <Element />
+        </FeatureFlags>
+      )
+    }
+    expect(HTMLRender(<Element />).container.firstChild).toHaveClass('test-class-name')
+    expect(HTMLRender(<FeatureFlagElement />).container.firstChild).toHaveClass('test-class-name')
   })
 
   it('should have no axe violations', async () => {
