+++ packages/react/src/__tests__/Avatar.test.tsx
@@ -1,9 +1,10 @@
 import React from 'react'
 import {Avatar} from '..'
 import theme from '../theme'
-import {px, render, behavesAsComponent, checkExports, expectRendersWithClassname} from '../utils/testing'
+import {px, render, behavesAsComponent, checkExports} from '../utils/testing'
 import {render as HTMLRender} from '@testing-library/react'
 import axe from 'axe-core'
+import {FeatureFlags} from '../FeatureFlags'
 
 describe('Avatar', () => {
   behavesAsComponent({Component: Avatar})
@@ -13,8 +14,22 @@ describe('Avatar', () => {
   })
 
   it('should support `className` on the outermost element', () => {
-    const element = <Avatar src="primer.png" className={'test-class-name'} />
-    expectRendersWithClassname(element, 'test-class-name')
+    const Element = () => <Avatar src="primer.png" className={'test-class-name'} />
+    const FeatureFlagElement = () => {
+      return (
+        <FeatureFlags
+          flags={{
+            primer_react_css_modules_team: true,
+            primer_react_css_modules_staff: true,
+            primer_react_css_modules_ga: true,
+          }}
+        >
+          <Element />
+        </FeatureFlags>
+      )
+    }
+    expect(HTMLRender(<Element />).container.firstChild).toHaveClass('test-class-name')
+    expect(HTMLRender(<FeatureFlagElement />).container.firstChild).toHaveClass('test-class-name')
   })
 
   it('should have no axe violations', async () => {
