+++ packages/react/src/__tests__/AvatarStack.test.tsx
@@ -1,8 +1,9 @@
 import React from 'react'
 import {AvatarStack} from '..'
-import {render, behavesAsComponent, checkExports, expectRendersWithClassname} from '../utils/testing'
+import {render, behavesAsComponent, checkExports} from '../utils/testing'
 import {render as HTMLRender} from '@testing-library/react'
 import axe from 'axe-core'
+import {FeatureFlags} from '../FeatureFlags'
 
 const avatarComp = (
   <AvatarStack>
@@ -34,7 +35,7 @@ describe('Avatar', () => {
   })
 
   it('should support `className` on the outermost element', () => {
-    const element = (
+    const Element = () => (
       <AvatarStack className={'test-class-name'}>
         <img src="https://avatars.githubusercontent.com/primer" alt="" />
         <img src="https://avatars.githubusercontent.com/github" alt="" />
@@ -42,7 +43,21 @@ describe('Avatar', () => {
         <img src="https://avatars.githubusercontent.com/github" alt="" />
       </AvatarStack>
     )
-    expectRendersWithClassname(element, 'test-class-name')
+    const FeatureFlagElement = () => {
+      return (
+        <FeatureFlags
+          flags={{
+            primer_react_css_modules_team: true,
+            primer_react_css_modules_staff: true,
+            primer_react_css_modules_ga: true,
+          }}
+        >
+          <Element />
+        </FeatureFlags>
+      )
+    }
+    expect(HTMLRender(<Element />).container.firstChild).toHaveClass('test-class-name')
+    expect(HTMLRender(<FeatureFlagElement />).container.firstChild).toHaveClass('test-class-name')
   })
 
   it('should have no axe violations', async () => {
