+++ packages/react/src/__tests__/Popover.test.tsx
@@ -1,9 +1,10 @@
 import React from 'react'
 import type {PopoverProps} from '../Popover'
 import Popover from '../Popover'
-import {render, behavesAsComponent, checkExports, expectRendersWithClassname} from '../utils/testing'
+import {render, behavesAsComponent, checkExports} from '../utils/testing'
 import {render as HTMLRender} from '@testing-library/react'
 import axe from 'axe-core'
+import {FeatureFlags} from '../FeatureFlags'
 
 const comp = (
   <Popover caret="top" open>
@@ -23,8 +24,22 @@ describe('Popover', () => {
   })
 
   it('should support `className` on the outermost element', () => {
-    const element = <Popover className={'test-class-name'}></Popover>
-    expectRendersWithClassname(element, 'test-class-name')
+    const Element = () => <Popover className={'test-class-name'}></Popover>
+    const FeatureFlagElement = () => {
+      return (
+        <FeatureFlags
+          flags={{
+            primer_react_css_modules_team: true,
+            primer_react_css_modules_staff: true,
+            primer_react_css_modules_ga: true,
+          }}
+        >
+          <Element />
+        </FeatureFlags>
+      )
+    }
+    expect(HTMLRender(<Element />).container.firstChild).toHaveClass('test-class-name')
+    expect(HTMLRender(<FeatureFlagElement />).container.firstChild).toHaveClass('test-class-name')
   })
 
   it('should have no axe violations', async () => {
